requirements.txt
anyio==3.6.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
contourpy==1.0.7
cycler==0.11.0
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==8.6.1
fairscale==0.4.4
fastapi==0.89.1
filelock==3.9.0
fonttools==4.38.0
h11==0.14.0
httptools==0.5.0
huggingface-hub==0.12.0
idna==3.4
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.6.3
nltk==3.8.1
numpy==1.24.1
packaging==23.0
pandas==1.5.3
parsedatetime==2.6
Pillow==9.4.0
pipreqs==0.4.11
pydantic==1.10.4
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.21.1
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sacremoses==0.0.53
sentry-sdk==1.14.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
timm==0.4.12
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tqdm==4.64.1
transformers==4.26.0
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
yarg==0.1.9
